# Optimize ProceduralSkyMaterial by removing uses of acos and simplifying logic