The changes can be clumped into three categories:
1. Remove use of `acos()`, instead do all calculations using cosines instead of angles
2. Only include the texture read from the cover texture if it is used. This was very expensive and accounts for about half the performance gains here _(edit: I optimized the calculation of `SKY_COORDS` so the performance gain from avoiding the cover texture will be reduced)_
3. Remove a bunch of unnecessary calculations by factoring them out or shifting them to the CPU. This made the code much easier to read, but the overall impact on performance was minor since a lot of the factorization was being done by the compiler. I suspect this will have a bigger difference on Android devices where the compiler doesn't optimize as aggressively. 

The result looks almost the same and runs over twice as fast.

### Performance ProceduralSkyMaterial
**M2 MBP**

I used a test scene with 3 4k Viewports that do nothing but render the sky. 
Before: 36 FPS (27.78 mspf)
After: 52 FPS (19.23 mspf)
No sky: 75 FPS (13.33 mspf)

Therefore we can say this reduced the cost of rendering the sky (to 4096\*4096\*3 pixels) from 14.45 ms to 5.9 ms! That's over twice as fast. Given that this is our default sky, I think that improvement is well worth making.

**Pixel 4**

This used a test scene with just a camera, 4 DirectionalLights, and a scaling factor of 2
Before: 44 FPS (22.73 mspf)
After: 52 FPS (19.23 mspf)
No sky: 60 FPS (locked) (16.67 mspf)

With resolution scaling that's rendering 2160 x 4560 pixels per frame with the sky. Saving 3.5 ms on the sky alone is pretty nice. At 1x resolution scale I am locked at 60 FPS in all cases

Looking at the GPU PROFILE when rendering at native res, this PR saves about 0.5 ms (of 3.0 ms). Not as much improvement as I expected

### Performance PhysicalSkyMaterial
**M2 MBP**

I used a test scene with 3 4k Viewports that do nothing but render the sky. 
Before: 46 FPS (21.74 mspf)
After: 53 FPS (18.87 mspf)
No sky: 75 FPS (13.33 mspf)

### Notes:

For posterity, this change was inspired by a blog post from Kostas Anagnostou above the [hidden cost of certain shader functions](https://interplayoflight.wordpress.com/2025/01/19/the-hidden-cost-of-shader-instructions/) (like acos, atan, etc.). Which led me to a lovely blog by Sebastian Lagarde on approximating acos/asin/atan to [drastically reduce instruction count](https://seblagarde.wordpress.com/2014/12/01/inverse-trigonometric-functions-gpu-optimization-for-amd-gcn-architecture/). I searched our codebase and started weeding out places where we use such expensive functions. ProceduralSkyMaterial is the worst offender by far. 

After replacing `acos()` in ProceduralSkyMaterial with an approximation (and seeing a nice performance boost) I realized that the shader could still be vastly improved and ended up removing `acos()` altogether. 

Similarly, I was able to remove `acos()` entirely from the PhysicalSkyMaterial, the places where it was used can be fully approximated with a linear curve instead without a noticeable change in performance. ([optical mass approximation desmos](https://www.desmos.com/calculator/v6g22btdc8), [sun energy approximation desmos](https://www.desmos.com/calculator/pkzsbg8xwx))

For PanoramaSkyMaterial (and any shader using `SKY_COORDS`) I relied on approximations of acos and atan so we would have identical behaviour, but improved performance.