@michaelharmonart 2025-01-24
love to see performance optimizations like this coming down the pipe! 

it might not be as flashy but those of us developing for tight performance targets like standalone VR really appreciate it!

(I remember profiling some scenes a while back in renderdoc and thinking the procedural sky render was taking oddly long for what I assumed it needed to actually do, so this is great news)