@clayjohn 2025-04-17
> Tested locally on Linux + NVIDIA, I noticed sky cover textures no longer appear to be seamless (even when the source texture is seamless):

Thanks for testing, I pushed up a quick change to the calculation of panorama coords last night that must have done this. I'll fix it quick. 

As for the sun rendering, the falloff curve for the glow is a little bit different since we are no longer using acos. The visual change is, IMO, acceptable for doubling the performance of the shader