@clayjohn 2025-04-18
I've made the suggested changes and optimized all three materials now! This PR should be ready for review/merging. 

Please note, there is a slight change in the appearance of ProceduralSkyMaterials (as you can see in Calinou's screenshots). The penumbra around the sun has a different falloff due to the switch from angle space to cosine space. I include an approximation so the falloff at default settings is very close. But at larger sun sizes / angles the curve gets further from the original. I think this is a fine tradeoff given the huge improvement to performance